requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0

# Database
sqlite-utils>=3.35

# AI
groq>=0.4.0

# Scheduling
apscheduler>=3.10.0

# Dashboard (Phase 3)
streamlit>=1.28.0

# Utility
python-dotenv>=1.0.0
pyyaml>=6.0.0
pandas>=2.0.0